+++ db/migrate/20161123204931_add_percent_to_sheets.rb
@@ -1,0 +1,6 @@
+class AddPercentToSheets < ActiveRecord::Migration[5.0]
+  def change
+    add_column :sheets, :percent, :integer
+    add_index :sheets, :percent
+  end
+end